flag for quick summary.
when working with very big collections, the generation of the summary
page becomes a real pain. but sometimes the user wants to take a quick
look from the summary output to see where they are in the process of
refining. the new flag, `--quick-summary` will generate a summary as
soon as possible with the least amount of essential information on bins.