A new function to learn about each nt
This function now allows us to say whether a given nucleotide position
in a partial gene call (a beginning or the end of which is missing from
the contig), a complete gene call, or that position is a third base in a
codon given the gene context. This will have pretty sweet implications
on our analysis results going forward. Related to #274. Closes #206.
Closes #163.