SO EXPERIMENTAL (and quite cool)
We no longer break conitgs from within genes to generate splits. I
haven't checked the performance of this code, although I don't expect it
to be anything but awesome (kidding). So, this heuristic will generate
very large splits if there are no places to cut them due to gene calls,
EVEN if you ask for very short splits. Probably addition of a
commandline flag for skipping this step is due. Related to #290.