source -> collection name
this is a part of a larger change and it will set the stage for the
ability to track the origin of each bin separately --whether they are
automatically identufued, or refined, or manually curated from the get
go. it was hard to add a new column to the table with the name 'source'
when source was already used for something it shouldn't have been used
for. change to 'collection name' resolves a lot of issues.