ability to track the source of a bin.
I apologize for the large, multi-file commit. with this commit we will
be able to track the source of each bin in a collection. we had a
problem with this: when a user refines an automatically identified bin
in a collection, there was no way to find out which bins were refined,
and which ones were not refined. Changes in this commit resolves that
issue once and for all.