a less I/O-demanding version.
initiating a transaction with the db for every gene call does not seem
to impact the I/O overhead that much when there is only one process, but
when multiple processes are running in parallel working with the same
file server, probably this becomes a very big burden. this should fix
it.