reimplement get_tree_object_in_newick (#372)
This one is tricky, but I have to explain it well. This function
converted tree objects generated by hcluster into newick strings. The
same tree objects are generated by SciPy, however it turns out that
those objects are not hashable. This reimplementation can work with
unhashahable objects. I will post a test to make sure the version of the
code for unhashable objects produce identical newick strings with the
version that works for hashable objects under #372.