new func determines default clustering ids (#377)
This is tricky. We have some deault clusterings defined in the
constants. For instance, for the merged profiles we want the
default to be 'tnf-cov', for single profiles we want it to be 'tnf',
etc. The problem is that these defaults do not indicate any
distance metric or linkages, even though anvi'o allows users
to define those variables freely in cluster configurations.

A clustering dict can contain multiple clustrings. The purpose
of this function is to take the desired default into
consideration, but then find a working one if it is not available,
or there are multiple ones in the dict.