some major changes for a more robust flow (#380).
now split names are directly coming from the profile db class, and we do
not need to do silly stuff to get split names that appear in the
profile database. also, the initialization of collection profiles are
done by passing only a collection id, and not a collection dict. bins
that contain no legitimate splits are not added to collection profile
which removes a lot of headache downstream.