import collections into profiles w/o contigs dbs
we were not able to import collections into a profile database if it was
blank, or simply not associated with a contigs databse as the program
required a contigs database parameter. now it is no longer a
requirement. but of course if the user is importing stuff into a profile
db without a contigs database they will have to be extra careful to make
sure names are consistent becasue anvi'o will have no way of checking
it.