a fix that closes #397. thank you Mike!
we were computing the average number of genes based on the number of
nucleotides found in a given contigs database. it is not a big deal if
one accesses to an external genome since the entire contigs db is
dedicated to a single genome in this case most of the time. but
accessing an internal genome identified by a bin id in a collection
stored in a profile db is a different story since the associated contigs
db contains nucleotides from all over the place (it is awesome that you
can make such silly mistakes that explaining the problem would take
much more characters than the fix itself).