introducing alphabets and contexts for HMM sources.
we now have a new file to describe the target of HMMs. should it be run
agains genes in AA alphabet? or genes in DNA or RNA alphabet. or contigs
in DNA alphabet, etc. each HMM source can require a different target,
and the way HMMs are handled now provides this flexibility.