py2 -> py3 fixes.
after running 2to3 with --write and --nobackups, I made about 30 changes
in various places to fix encoding issues. currently the anvi-self-test
is failing due to the fact that new profiler does not generate
RUNINFO.cp files. although, anvi-self-test makes it all the way to the
anvo-merge successfully, which is a good start. the next bunch of
commits will focus on RUNINFO.cp issues.