this fill fix this issue: https://goo.gl/hycFqI
when we use external gene calls, AA sequences are not coming from
Prodigal, but we compute them based on the DNA seqeunce we find in ORFs.
clearly we were adding * characters for STOP codons to obey the common
practice, but this is not compatible with the output the rest of the
pipeline is used to.