attempting to edit anvi-search-functions-in-splits
I have helped a couple people who wanted to search for functions by accession ID (like PFXXXXX/COGXXXX). I usually hack around by uploading another functions table where the accession ID is in the 'function' column also. I tried to propose a change here that would search the 'accession' column would be searched too, but i'm not sure of 2 things: 1) if you didn't do it this way for a reason; and B) if this is the correct or only thing that should be changed to actually add what i'm talking about

hopefully that makes sense!
-mike