add more gene coverage stats things to summarizer.Bin
Similar to 4df096431efa12518c1a6ff8ddb9fe94c6bd769b, but now adding it
to the bin. This is probably a poor design, but without this addition,
we wouldn't be able to use the mcg classifier on a collection, since we
need access to these gene coverage stats objects.
@meren, please look at this change and let me know if you think it is
ok.