slight change in how arguments are parsed.
this way we can put in flags ad hoc. probably a better way would be to
do the parsing elsewhere, and look into the 'unknown' to make sure there
is only --debug there and give user an error if they are trying to use a
flag that is not in fact there.