added transpose option now minnie test works
re #663.
Notice that we were getting the error because we weren't transposing
the data. The actual error generated (`KeyError: 'order_type'`) is
generated while the an exception is trying to be generated. I don't
know why there was no key "order type". Maybe this error handling
should be changed. I'm just putting this note here because I'm not sure
what was the deal with the missing key.