improvements in contigs
so it knows when it is inherited from within another class, or inherited directly. if it is inherited from within another class, it expects no `args` to be passed. if it is inerited directly, then the user must pass the `args`. also, if the inheriting class does not have a `name` set, it is a problem.