added input arguments: `--list-dependencies` `--config-file`
@meren, Sadly, your changes broke the snakefile and it can't run now.
This is due to the fact that I didn't yet implement the test so you
couldn't test it...

I couldn't get to fix it now with the limited time I had. If you want,
the following command should work:
```
snakemake -s
/Users/alonshaiber/github/anvio/anvio/workflows/contigs/Snakefile
--configfile /Users/alonshaiber/Downloads/config.json -np
```

But it doesn't right now. You can ignore this message, and I will fix
it sometime when I get to it, either this weekend or next week.