only check config if we run anvi-run-snakemake-workflow
This is very important. This means that if someone will use the
snakefile as a standalone (not calling it through the
anvi-run-snakemake-workflow program) then there will be no examination
of the config.

The reason I had to make this change is because of the include
statement, when we include the contigs workflow snakefile in the
pangenomic workflow snakefile, if the test for the config was included
in the contigs workflow, then it would raise an error because the
config includes pangenomics parameters which are not familiar to
contigs.