samples.txt changed to fasta.txt in pangenomics workflow
in order to be consistent with contigs workflow language. Maybe we
should just call this file `contigs.txt`, we will decide later, as this
is an easy change. The most important thing is that it HAS to be
consistent in contigs workflow and in all inheriting workflows (i.e.
both metagenomics and pangenomics)