added rule import_percent_of_reads_mapped to metagenomics workflow
This will be imported to the individual profile
In order to make this happen, I added the flag as an input to
anvi-merge, so that merge would only happen if the percent of reads
mapped was imported.
The user can choose not to import this (if for example they didn't save
the log files from bowtie, and they want to avoid an error...). The
function `input_for_anvi_merge` is the one that figures out whether to
run ` import_percent_of_reads_mapped` or not.