fix for get_fastq in metagenomics workflow
now get_fastq is consistent in that it always returns the fastq files
for a single sample, whether the user is running qc or not.
Sadly now we have each of the following idba_ud, megahit and bowtie
access the fastq files in three different ways, but for now I didn't
invest the time to fix that.