bypassing weird snakemake bug
Ok, so sit tight, because this change shouldn't really change anything.
But it does...

The essence of this change is to avoid using `unpack`, and hence there
is a separate function to configure the input for profiles and the
input for `percent_of_reads_mapped_imported_flag`.

I'm documenting all the details here in case I will get the chance to
reproduce this and report to the snakemake developers:

From some reason if we use `unpack` and then later pass
`input.profiles` as input to the function
`remove_empty_profile_databases`, then things sometimes work and
sometimes don't. And the weirdest thing is that they don't work if the
workflow has never been run before (OR if you delete the `.snakemake`
directory), but if after it fails once you try it again then it works.
So this seems to me like there is a `type` inconsistency between whther
snakemake is using the things in `.snakemake` to make sense of input
files, or whether unpack is used.

The part that fails is inside the function
`remove_empty_profile_databases`:

```
    for p in profiles:
        db = dbops.ProfileDatabase(p)
        n = next(iter(db.meta['total_reads_mapped'].values()))
```
from some reason, if I print `profiles` then it looks fine, but then if
I print `p` inside the for loop then it sometimes prints `0` instead of
the expected string (the path to a profile database).