unzipping fastq files if they are compressed and if no QC is done
This important commit allows us to use idba_ud without performing QC
and without keeping the fastq files uncompressed. This is good if we
are assembling fastq files for which QC was performed already.

The downside of this implementation is not compressing and
uncompressing is done sequentially for each member of the pair.

The alternative would have been to write two more rules one for
compressing and one for uncompressing the input to fq2fa. I chose the
current solution to increase readability and reduce complexity of the
code. we can always change this if it proves to be annoying.

I wonder if anyone would ever read this commit message. if you do,
please let me know, it would make me happy, but also sad that I have
wasted your time. Thank you very much