computing functions enrichment in groups
@meren, this is currently producing a dictionary with three levels,
where the keys are category name, function, name, enrichment score,
respectively.
There are two enrichment scores, `"enrichment"`, and
`"weighted_enrichment"`, the former is the straight forward one that we
discussed, and the latter is a weighted version of that score.
The purpose of the weighted one, is that if you would want to compare
the enrichment scores across all groups, then you would need to account
for the difference in group size (I chose the first idea that came to
my mind, so this could be complete bulls$%t).

I'm not creating the output yet, because I'm tired, and thought I would
give you a chance to take a look, and tell me what you think :-)

I hope you are having a good night!