anvi-run-snakemake-workflow -> anv-run-workflow
Originally we wanted to keep snakemake in the name of the program so as
to give snakemake the credit it deserves. While we still agree with
this notion, we feel that this is making the name of the program tool
long, and that we can give snakemake it's credit (and we do) in other
ways (for example, see #854, and also in the tutorial and help menu).