fixy.
this is a bit hard to describe.

when there is a collection with a single bin that describes all splits, this line turns splits of interest to None, and it causes downstream issues.

probably at some point we will have to deal with performance issues that may arise from the removal of this line though.

related #883.