address COG database issues
This is related to #894.

It turns out changes we made to the pangenomics workflow awhile ago had unintended effects on how COG setup module worked. As a result, anvi'o was unable to setup COG search databases properly. These changes solves that while not breaking the pangenomic worklfow.

I apologize for the previous error.