sanity_check_for_args -> init_workflow_super_class
only if nothing was initiated then we expect a config and then we also
initiate the WorkflowSuperClass.
This probably means that we should only have the WorkflowSuperClass
expect and assign the config file, and then the design would make more
sense. I will think about it, but for now, this works, and a proper
default config file is generated for MetaPanGenomeWorkflow.