add some run_virsorter rule params, and set default to NOT run virsorter
@jarrodscott, notice how I set the default. You can follow this later
on, to set more default parameters (for example the number of threads.
Also, notice line 91. In order for a parameter to be available for
change through the config file, it must be defined in this way. This
should include also things like this:
```
            dataset = M.get_rule_param("run_virsorter", "--dataset")
```

I left it for you to add the rest of these, if that's ok.