Added virsorter to self.default_config.update
For `import_misc_data_virsorter` I set the default of `--target-data-table` to `items`. This is mandatory and should not be changed. One thing I was not sure about was setting the default collection name for `import_collection_virsorter`. Will avni'o fail if no collection name is given? If so, we can provide a default name here but maybe a user would want to change the default value. How can this be implemented?