new default thread numbers for rules
I reduced the number of threads to 1 for those rules that don't really utilize multi-threads. These were originally greater than 1 just to reduce the number of things that could run in parallel, but now we  have "resources" to control that.

I also reduced the default number of threads utilized by those rules that actually use threed.

@meren, this is not too critical, but let me know if you think any of these numbers should be different. Thanks!