Getting rid of the extra bowtie-related rules
This is one of my favorite commits ever.
I moved get_fasta and get_raw_fasta to the backhand, where they have basic functionality in the ContigsWorkflow backhand, and it is extended by MetagenomicsWorkflow class.

Now we have a fasta_information dictionary that holds the information both for references and for references-for-removal.

contigs_information is replacing the old fasta_information, and holds only information for the references.