get rid of intermediate fastq files
We accumulate a lot of intermediate fastq file, and haven't been doing a good job of deleting them.

Here I make sure that if a fastq.gz file is uncompressed then the uncompressed copy is deleted.
Also, I make sure that that the output of QC is removed if short read filtering was done (since there is no point in keeping both, as the output of QC is the input to the short read removal, and if you regret removing reads you can just go back to your raw fastq file, and repeat QC. Also if you just want the sequence of the reads that were removed you can use the id list and get these from the raw fastq file.