fix for get_raw_fasta
When gunzip_fasta rule calls get_raw_fasta in the metagenomics workflow, then it refers to the MetagenomicsWorkflow get_raw_fasta and not the ContigsWorkflow get_raw_fasta, so we need to make sure we call the super class function if we are dealing with a reference for removal.

In addition, I also got rid of the no-longer existing rules.