misc data targets genes db args calls for gene mode
but if the gene database is not there desite the args.gene_mode, then throws a specital error, which is going to be handled elsewhere.

the design flaw here is that the interactive interface wants to read misc data before it initializes the genes database. otherwise this exception here wouldn't have been necessary.

in fact, in an ideal world it shouldn't be here.