if you get a genes db error, omit it.
this is a design flaw.

I am not sure why interactive has to recover aditional data items and all here. but this creates a catch 22.

in gene mode, anvi'o wants to read from misc data tables of the genes database.

but in genes mode, there may be no genes database yet, as it is generated automatically later in the code.

when we try to read from an non-existent db it should cause issues, but not in this case. so this workaround tries to deal with that.