added anvi_split and better configuration of target files
We should also do a similar change for other workflows, but for now it's just the metagenomics workflow.
With this commit the metagenomics workflow will have it's target files (i.e. the input for the target rule: metagenomics_workflow_target_rule) are configured in the backhand.

In addition, in this commit I added the "run" parameter for anvi_summarize so people can choose whether to run it or not (assuming that they provided a collection_txt file).

similar to anvi_summarize, if anvi_split is configured to run, it will only run on groups that are included in the collection_txt file.