include tree name in import tree flag
this way we can run this multiple times with different config files and get multiple trees imported.
Also, from here on the sequences will be imported into the phylogenomics directory (PHYLO_DIR), because it makes more sense, and also will help with the case of importing multiple trees.

So in order to import multiple trees one must give each tree a unique name, and also provide a new name for the PHYLO_DIR for each tree.