fixy fix. report HMMs properly even in mock cases
this was the most convoluted bug I’ve ever run into in anvi’o, and at the heart of it lies the fact that it is only triggered when using the same contigs to create multiple contigs databases and then using anvi-get-seqeunces-for-hmm-hits program. there is a step where we generate unique ids for every gene / hmm hit in every contigs prior to writing hmm results to the database to make sure we don’t get confused for HMM hits that are in multiple splits but meant for the same gene (due to soft spligging). when the same unique id is assigned to a gene/hmm hit in different contigs databases, and they are used as external genomes, shit hits the fan. so even this level of specificity of the unique ids is not enough:

```hit['gene_unique_identifier'] = hashlib.sha224('_'.join([gene_call['contig'], hit['gene_name'], str(gene_call['start']), str(gene_call['stop'])]).encode('utf-8')).hexdigest()```

for an hmm hit to have the same gene_unique_identifier in two different contigs databases, the contig name, gene name, and gene start/stop positions must be identical. my mistake here was to not include something that is ABSOLUTELY unique to a contigs database, like the hash. I did it now, and even the most ridiculous test cases with identical contigs databases will work :)

yay.