dont create fake profile database when group is one.
closes #1068.

As of this commit, if the user specified groups of size one, then in the merged directory, under a directory with the group name, a README.txt file will be created and it will direct users to the individual profile database.

Notice that if a group had more than one sample, but only one (or none) of the samples recruited any reads (i.e. all bam files were empty), then to avoid raising an error I'm still creating fake profile databases (but I'm also writing a message to the log file. We can change this behaviour if @ozcan, or @meren think this is not favorable.