Allow to skip reformat per fasta file in fasta.txt
@meren, in order to be able to import gene calls to genebank genomes, we can

1. either skip reformat fasta
2. or reformat the fasta and then translate the names of contigs in the gene calls table you create with anvi-script-process-genbank-metadata to the new names of contigs

To allow for option 1, in this commit I changed how we handle skipping reformat fasta. So far, if you want to skip reformat then you set `run` to false inside the params for the rule `anvi_script_reformat_fasta`. Instead (and I think this is better) as of this commit the `run` parameter will no longer be available (rendering many existing config files as unusable), and instead you can add a column in your `fasta_txt` `skip_reformat_fasta`, and for each entry that you put `"1"`, then we skip reformatting.

I wanted to ask you whether you think I should still allow users to use the `run` parameter, so that if it is set to `false`, then reformat is skipped for all entries in your `fasta.txt` file? What do you think? I'm not a fan of this because I don't like it when there are multiple ways to set the same thing.

One last thing:
Next, I plan to make it so that it would be possible to run reformat when supplying gene calls, and I will add a step to translate the gene calls. The reason for this is that it standardizes things so that we can expect all contigs to have the prefix as defined in the `fasta_txt` file. This will be particularly useful once I finish the metapangenomic workflow, since in that workflow we concatenate all fasta files, and if we control contig names then we can make sure that two are not alike.