the output of the second step of reformat is temporary
The reason for why we have two steps of reformat is explained here: #926. And there is no reason to keep two copies of nearly identical or sometimes identical fasta.

Also from now on we generate a report (temp file also) for this step.
We will only use this report if external gene calls or external functions were supplied.