the output of the second step of reformat is temporary
The reason for why we have two steps of reformat is explained here: #926. And there is no reason to keep two copies of nearly identical or sometimes identical fasta.

We don't create a report for the second step since the report doesn't track removed contigs, it only tracks the renaming.