dont raise error for missing dependencies
There will be a warning that dependencies may be missing, but no point in raising an error since the user's configuration may lead to avoid needing some dependencies. Once we make a better solution that actually takes the DAG into consideration, then we can set this to raise an error if a relevant dependency is missing.