if there are domains in the training directory, give the user a warning. re: #1154
Unfortunately anvi'o will require any new, user-defined SCG collection to be stored in the default anvio/data/hmm dir as a proper directory. this is kind of frustrating, but it is necessary at this point to not invest a complete rewrite effort.

I will put together some material for users to be able to re-train the RF classifier with their new domains soon, and update the code and messages with links to that resource.