using gc identities as functions as an option
this enables functional enrichment analyses to be done on gene cluster
level, instead of function names that are assigned by COGs or EggNOGs or
w/e. Special thanks go to Sam Minot for bringing this up.

Using this will require the inclusion of the flag `--include-gc-identity-as-function`,
which will add a new annotation source that assigns each gene cluster to
its name, as well as the use of the parameter --annotation-source 'IDENTITY',
to rely on that new information.