even more room for error.
This another fix for @jarrod. Anvi'o throws this error if it can't find a PFAM accession ID in PFAM catalogue:

```
Config Error: It seems hmmscan found a accession id that does not exists in Pfam catalog, Id: PF00001
```

With this change, anvi'o will not crash, but add a function for this entry called "Unknown function with PFAM accession X" where X represents the PFAM id for which the catalogue does not have any informaiton.