change in design. DIAMOND databases are generated at the user-side
we quickly realized that changes in DIAMOND version between computers break everything. the new design separates the download and preparation of GTDB files from the generation of databases. this way we can ship all the FASTA files, but the user can create databases on their computer without having to create a burden for the GTDB team.

this is related to issue #1240.