add self.profile_databases to MetagenomicsWorkflow class
This is addressed at dealing with single profiles (i.e. groups of size 1), vs. merged profiles.

The reason this need is rising now is due to the ability to import collections and summarized. if this is a group of size 1, then we need to know that the collection needs to be imported to the single PROFILE database.