require snakemake 5.7.0
I just realized I have been using 5.7.0 on my local as I test the workflows.

Two users recently complained about bugs and I realized that the bugs are solved by upgrading from Snakemake 5.2.4 to 5.7.0. I don't understand why, but it is what it is.

Here are the bug reports and conversations:
https://anvio.slack.com/archives/C8SFMGYF3/p1576263544012400

https://anvio.slack.com/archives/C8SFMGYF3/p1576262243010600

Thank you @rbartelme and @jarrodscott, for reporting!