don't report codons with ambiguous nucleotides
this is a problem brought to our attention by Ben Tully. the function `get_list_of_codons_for_gene_call` returns None when a codon is `constants.codon_to_codon_RC` (i.e., 'ANN'). but then this only a problem for genes that are reversed. `constants.codon_to_codon_RC` is not used for forward genes, and codons are reported with ambiguous characters.

this solves that issue by *not* reporting at all any codons with ambiguous bases. if we need to change this behavior we can add a flag to this function later on, but in that case we will have to do a smarter version of this logic to get reverse complemetns:

```
constants.codon_to_codon_RC[codon_seq]
```