update contigs db version.
we don't have a migration script at the moment, but @ekiefl promised for one :p

the key there is to add a call_type column which, for every gene that has a non-empty entry in the amino acid sequence table, has a value of 1, and 2 for the rest.

plus we have three new self table entires (since we will not know at the time of migration whether the user has used an external genes file, we will assume all to be prodigal with defaults).