"internal" goes away and "provided_by_{anvio|user}" comes.
this way we can better describe the properties of an artifact.

some artifacs (i..e, state-json) can be generated and reported by anvi'o, but can also provided by the user. in contrast, oligotypes will only be reported by anvi'o, and external genomes will only provided by the user. the "internal" attribute alone was not helping for a comprehensive description.