SCG taxonoy ops now works with a compressed file, rather than a flat text file, for accession id to taxonomy.
the original idea was to keep this as a flat text file so we could follow the progress of addition and deletion of genomes. but keeping 120,000+ lines of text in git is not really a very good idea :/ I realized this while getting ready to start working on tRNA taxonomy, and wanted to fix it once and for all.