use `project_name` rather than file path as `name`
when the contigs workflow inherited by pangenomic or phylogenomic worfkflows, it generates an external genomes file and sets the names of contigs databases based their path name. this changes that behavior and enforces the use of `project_name` in the self table to be used as name (which will remove all the `*-contigs` extensions to names in all pangenomes going forward).