set the window title for inpsection page, and thus solve ONE OF THE MOST ANNOYING THINGS IN ANVI'O.
it is astonishing that the anvi'o inspection page is one of the most frequently used features of anvi'o, but at no point during the last few years we stopped and said "well, perhaps http://0.0.0.0:8080/app/charts.html?id=longasssplitname&show_snvs=true is not the best title to show on the browser window" :joy:

I did it this Sunday. and it will be in v7. You're welcome, future me.