fixy fix that closes #1565.
I also would like to admit that it took 4 hours to find this. essentially what happens is this: we get a copy of gene call from the self variable, and then modify it downstream, which modifies the self. when we are back to this function, now everything is ruined and you have no idea why ever gene call of yours start at position 0 of your damn contigs.