self.contig_sequences dictionary may not be empty, but it may also be missing sequences for the new set of genes of interest.
here is what happens. the first time you call this function, it initializes contig sequences for a given gene call. then you call it again using the same ContigsSuper instance with a gene call that is in another contig. the if statement says "I have my contig sequences" and moves on, but downstream code explodes because the new gene caller id is on another contig that was not initialized. with the new variable we can see if the new set of gene calls of interest are among those that occur in the contigs that are already initialized.

bring me bugs like this. difficult, but fun to solve. don't bring me bugs due to some memory black magic that sucks your life out of you while invoking "I wanted to go to kitchen to drink some water at 3am and the next thing I know is that I am here walking with you guys to rob a bank in TexaWHO ARE YOU PEOPLE AND HOW COME I DON'T EVEN HAVE A MASK MY WIFE WILL KILL ME"-type confusion.