This commit comments out some lines related to downloading pathways.
We are not using pathways yet downstream, so there is no need for users to download them.
However, we will use them at some point in the future, and when that happens, this commit
can simply be reverted to go back to a state at which pathway downloading is working.