anvi-migrate provides nothing.
yes, it actually provides something, but a program should never provide what it requires.

what to do, then? a similar situation happens with `anvi-init-bam`. the program takes a BAM file and produces a BAM file. to solve this, we have two artifact types: RAW-BAM and BAM. so it takes the former, and produces the latter. in a simliar fashion we would need other artifacts such as contigs-db-outdated, profile-db-outdated, and so on. but then it would complicate things way too much in the network. there are better solutions, but they are technically demanding and not quite necessary (such as turning our artifacts into nested artifacts, where all databaes could be also defined as "anvio-db" in which case we can have a single "outdated-anvio-db", but it is not really necessary at this point). so the next best solution is to make sure anvi-migrate does not provide anything :)