an extremely important detail re: h5py
it turns out installing newer versions of h5py totall breaks our
migration of databases from a few years ago. while it is recoverable
with no problem, it would have been a horrible experience for the users.

testing the release candidate is pretty boring, but it reveals a lot of
minor and major issues.