a new variable for `self` tables: db_variant
although it is only used for contigs databases at the moement, we can
expand its utility in the future. The need for this comes from the fact
that our contigs databases can now serve multiple purposes, and giving
the user/programmer option to set explicitly what variant they are can
empower us to be able to do more things downstream without having to
predict what a contigs db is. the first applicaiton of this will be to
distinguish trnaseq databases from everything else.