updated the error message because this was confusing me
it is possible but probably rare to produce input with no associated 
groups (I was testing on two copies of the same genome in different 
groups), but I was confused because I thought it was talking about my 
input groups file (which does have more than one group). hopefully this 
clarifies things. and probably if there are no associated groups in your 
input file we should fail even before we get to this script