this is an important change
we report sequences for contigs and genes using a dictionary type object
where keys are contig names or gene caller ids, and items are
dictionaries that look like this: {'sequence', seq}. this is not the
case for the way we report split sequences in dbops, which cause some
design headaches. this makes the reporting of split sequnces identical
to contig and gene sequences, and fixes the usage in multiple places.