for detection view, load full state by default.
this change makes sure that the detection view is loaded with the global
state values in refine mode. in refine mode, we avoid using the global
state values for views since global values often poorly serve for
indiviudal bins. but in the case of detection, it is never true and not
using the global defaults create confusing situations. this changes
fixes this problem.