GenomeView class
the purpose of this class is to collect everything necessary for the
interactive interface to read through bottleroutes. It is likely that we
will need a way to re-implement this class to be lazy (so it doesn't
read everything into memory and pass it around). but this draft will
serve well to figure out how to structure the interactive interface with
real data.