fix #1718
turns out pandas.read_csv is NOT flexible to read files where some lines 
have more cols than given in the 'names' param. so instead we now turn 
the file into a list of lists of inconsistent length and make the df 
from that list, which does work. and we can generate an appropriate 
number of column names afterwards.