key change in Amy's script
we realized that `associated_groups` column does not have to include
more than 1 group names. there may be only one group in that column, and
things would still work. for instance, if one has two groups (A and B), and one
group (A) includes a set of genomes that are a subset of the genomes in the
other group (B), no function will be associated with the group A, while
there may be functions asscoiated with group B. In which case,
`associated_groups` column will only have B. The older version of the
script gave an error here. This commit fixes that behavior.