now we have no __parent__ in single profiles
which means, our view data are a bit lighter, yet missing a key piece of
information that is necessary to show which splits belong to which
contigs. the BEST place to do this in fact is the JavaScript side of
things to avoid redundancy of data in views dict. but unfortunately
doing it here is the most convenient. so here we are adding the missing
'__parent__' layer to each view data.