a new table form for clusterconfigs
since our view tables are not MxN matrices anymore, and they don't fit
the definition of dataframes as they are defined in clustering module,
here I add a new definition, 'views', so the view data can be stored
in MxN matrix form without any trouble. I wrote and removed 100 lines of
code before I realized that this could be handled like this. in the
depths of the anvi'o codebase no one even looks in, there are some
design gems :p