NO MORE relative_abundance or max_normalized_ratio
This is very important. I am removing all the code from the merger
module that calculates normalized coverage values for splits across
samples.

I am doing this because I don't think there is any way to do this right,
and I don't think anyone benefits from having those in merged profile
databases. if we believe there are things we could do better here, we
will need to reimplement them in a smarter way. the current design was
causing a signficinat bottleneck in this module for many things.