no more atomic data table for single profiles.
now, single profiles and merged profiles are structurally identical,
which brings multiple significant changes to the design all over the
codebase from clustering configurations to the computation and storage
of profile 'view' data, to summary and splitting operations.

While this commit does not cover all of the necessary changes, it puts
together a blueprint to follow.

The next critical task is to sort out `anvi-summary`, `anvi-split`, and
the default visualization of single profiles.