READ FULL COMMIT MESSAGE FOR DETAILS.
This is a very quick fix that speeds up coverage accession, reduces
auxiliary-db size, and even speeds up anvi-profile slightly.

Here's the idea. It's often the case that a split will yield 0 reads,
resulting in large numpy arrays of zeros. This is especially true of
inversion profiles and large assemblies. Why waste our time storing
these? Not only is it a waste of space, but querying the DB becomes
slow due to decompressing coverage arrays.

This simple solution is to determine whether a split's coverage is all
zeros. If not, then we do what we've always done, we store a gzipped
byte representation of the coverage array. But if so, we instead store a
simple integer number, which represents the length of the split. Then,
when querying split coverages, we either decompress the array if the
blob in the table is a bytes object (like before), or we construct
a zeros array with np.zeros(...) if the blob is an integer.

To test out this new change, I profiled a sample that yielded low
recruitment to the FMT assembly. In this case, 85% of splits had all
zero coverage. Profiling was about 2% faster, the auxiliary-db was 30%
smaller, and coverage accession was 90% faster.

This change does not require a migration script.

NOTE: A more ambitious idea was first tried, in which I created a
'sparse coverage representation' vector. The idea was to take compress
stretches of zeros by starting with a standard coverage array like this:

[0,0,0,5,23,36,33,22,1,0,0,0,0,0,5,5,6]

and compressing them to this:

[-3,5,23,36,33,22,1,-5,5,5,6]

where negative values refer to the length of a stretch of zeros. I stand
by the fact that this is a rather smart idea, however my implementation
that converted between sparse arrays and coverage arrays ended up being
too slow to be worth it, even with numba implementations. It also
required changing the dtype from uint16 to int32, which kind of sucked.